feat: cambio rutas udistrital/planeacion_cliente/issues/972
